Infering of django model fields is moved to a dedicated module.
